feat(agent): add shared timed window helper; drivehealth uses it
Extract the local `timed<T>` acquisition-window helper from drivehealth
into `src/agent/timing.rs` (Linux-gated) so other Regime-R samplers can
reuse it; update drivehealth/linux/device.rs to import from the shared
location and delete the now-redundant local copy and its test.
